Create 24th(r RSA system
RSA ENCRYPTOR/DECRYPTOR
*******************
PLEASE ENTER THE 'p' AND 'q' VALUES BELOW:
Enter a prime number for p: 2
Enter a prime number for q: 5
*******************
RSA Modulus(n) is: 10
Eulers Toitent(r) is: 4
*******************
The value of e is: 999
*******************
EUCLID'S ALGORITHM:
4 = 0*(999) + 4
999 = 249*(4) + 3
4 = 1*(3) + 1
END OF THE STEPS USED TO ACHIEVE EUCLID'S ALGORITHM.
*******************
EUCLID'S EXTENDED ALGORITHM:
1 = 4*(1) + (-1)*(3)
1 = 999*(-1) + (250)*(4)
s=-1. Since -1 is less than 0, s = s(modr), i.e., s=3.
END OF THE STEPS USED TO ACHIEVE THE VALUE OF 'd'.
The value of d is: 3
*******************
Private Key is: (3, 10)
Public Key is: (999, 10)
*******************
What would you like encrypted or decrypted?(Separate numbers with ',' for decryption):2
Your message is: 2
Type '1' for encryption and '2' for decrytion.1
Your encrypted message is: []
Thank you for using the RSA Encryptor. Goodbye!